Renamed view files for naming consistency
