+++ rampp2p/views/__init__.py
@@ -1,9 +1,9 @@
-from .views_utils import *
-from .views_webhooks import *
-from .views_user import *
-from .views_ad import *
-from .views_order import *
-from .views_contract import *
-from .views_appeal import *
-from .views_payment import *
-from .views_misc import *
+from .view_utils import *
+from .view_webhooks import *
+from .view_user import *
+from .view_ad import *
+from .view_order import *
+from .view_contract import *
+from .view_appeal import *
+from .view_payment import *
+from .view_misc import *
