+++ rampp2p/views/view_ad.py
