+++ rampp2p/views/view_contract.py
